+++ nop-rule/nop-rule-core/src/main/java/io/nop/rule/core/execute/RuleRuntime.java
@@ -1,3 +1,10 @@
+/**
+ * Copyright (c) 2017-2023 Nop Platform. All rights reserved.
+ * Author: canonical_entropy@163.com
+ * Blog:   https://www.zhihu.com/people/canonical-entropy
+ * Gitee:  https://gitee.com/canonical-entropy/nop-chaos
+ * Github: https://github.com/entropy-cloud/nop-chaos
+ */
 package io.nop.rule.core.execute;
 
 import io.nop.api.core.time.CoreMetrics;
@@ -33,7 +40,7 @@ public class RuleRuntime implements IRuleRuntime {
 
     private String ruleName;
 
-    private Integer ruleVersion;
+    private Long ruleVersion;
     private boolean ruleMatch;
 
     private boolean collectLogMessage;
@@ -87,12 +94,12 @@ public class RuleRuntime implements IRuleRuntime {
     }
 
     @Override
-    public Integer getRuleVersion() {
+    public Long getRuleVersion() {
         return ruleVersion;
     }
 
     @Override
-    public void setRuleVersion(Integer ruleVersion) {
+    public void setRuleVersion(Long ruleVersion) {
         this.ruleVersion = ruleVersion;
     }
